Fix storing normalized cookie with aiohttp 3.10
aiohttp 3.10 and more specifically https://github.com/aio-libs/aiohttp/pull/7944 changed the way cookies are handled. Previously, the domain did not have to be part of the cookie jar for a cookie to be used. As this was changed in 3.10, we now need to pass in the response URL when updating the cookie jar, otherwise the domain will not be part of the stored cookie, and consequently will not be used for the next request.

Fixes https://github.com/home-assistant/core/issues/123310.